bug report by Adriano dos Santos Fernandes also applied to InstanceQuery (SHORT got read as BYTE) https://list.scms.waikato.ac.nz/mailman/htdig/wekalist/2008-September/014332.html
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/branches/book2ndEd-branch@4516 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ weka/core/converters/DatabaseLoader.java
@@ -66,7 +66,7 @@ import java.util.Vector;
  * Sets incremental loading
  *
  * @author Stefan Mutter (mutter@cs.waikato.ac.nz)
- * @version $Revision: 1.1.2.4 $
+ * @version $Revision: 1.1.2.5 $
  * @see Loader
  */
 public class DatabaseLoader extends AbstractLoader implements BatchConverter, IncrementalConverter, DatabaseConverter, OptionHandler {
@@ -899,7 +899,7 @@ public class DatabaseLoader extends AbstractLoader implements BatchConverter, In
 	  }
 	  break;
 	case SHORT:
-	  short sh = rs.getByte(i);
+	  short sh = rs.getShort(i);
 	  if (rs.wasNull()) {
 	    vals[i - 1] = Instance.missingValue();
 	  } else {
@@ -1059,7 +1059,7 @@ public class DatabaseLoader extends AbstractLoader implements BatchConverter, In
 	  }
 	  break;
 	case SHORT:
-	  short sh = rs.getByte(i);
+	  short sh = rs.getShort(i);
 	  if (rs.wasNull()) {
 	    vals[i - 1] = Instance.missingValue();
 	  } else {
