+++ weka/experiment/InstanceQuery.java
@@ -47,7 +47,7 @@ import weka.core.*;
  * Command line use just outputs the instances to System.out.
  *
  * @author Len Trigg (trigg@cs.waikato.ac.nz)
- * @version $Revision: 1.15.2.1 $
+ * @version $Revision: 1.15.2.2 $
  */
 public class InstanceQuery extends DatabaseUtils implements OptionHandler {
 
@@ -375,7 +375,7 @@ public class InstanceQuery extends DatabaseUtils implements OptionHandler {
 	  }
 	  break;
 	case SHORT:
-	  short sh = rs.getByte(i);
+	  short sh = rs.getShort(i);
 	  if (rs.wasNull()) {
 	    vals[i - 1] = Instance.missingValue();
 	  } else {
